requirements-docker.txt
fastapi
uvicorn
tensorflow-cpu
numpy
pillow
python-multipart
scikit-learn
matplotlib
streamlit
requests
pandas
